actualizacion de read-me
