+++ internal/application/src/algorithms/simple_linked_list_applications/HighPrecisionIntegerOps.go
@@ -1,7 +1,6 @@
 package simple_linked_list_applications
 
 type HighPrecisionIntegerOps interface {
-	SetUpHighPrecisionInteger(number int) HighPrecisionInteger
-	Sum(number1 HighPrecisionInteger, number2 HighPrecisionInteger) HighPrecisionInteger
-	Difference(number1 HighPrecisionInteger, number2 HighPrecisionInteger) HighPrecisionInteger
+	Sum(number2 HighPrecisionInteger) HighPrecisionInteger
+	Difference(number2 HighPrecisionInteger) HighPrecisionInteger
 }
