+++ internal/application/src/algorithms/simple_linked_list_applications/HighPrecisionIntegerOpsImplemented.go
@@ -1,0 +1,9 @@
+package simple_linked_list_applications
+
+func (hpi *HighPrecisionInteger) Sum(number2 HighPrecisionInteger) HighPrecisionInteger {
+	return SetUpHighPrecisionInteger(1)
+}
+
+func (hpi *HighPrecisionInteger) Difference(number2 HighPrecisionInteger) HighPrecisionInteger {
+	return SetUpHighPrecisionInteger(1)
+}
